requirements.txt
PyMySQL==1.1.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
mysqlclient==2.2.7
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1